query planner join list
