+++ hopsworks-api/src/main/java/io/hops/hopsworks/api/featurestore/FeaturestoreService.java
@@ -16,6 +16,7 @@
 
 package io.hops.hopsworks.api.featurestore;
 
+import com.google.common.base.Strings;
 import io.hops.hopsworks.api.featurestore.featuregroup.FeaturegroupService;
 import io.hops.hopsworks.api.featurestore.storageconnector.FeaturestoreStorageConnectorService;
 import io.hops.hopsworks.api.featurestore.trainingdataset.TrainingDatasetService;
@@ -59,7 +60,6 @@ import io.hops.hopsworks.restutils.RESTCodes;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import org.apache.parquet.Strings;
 
 import javax.ejb.EJB;
 import javax.ejb.TransactionAttribute;
@@ -453,5 +453,16 @@ public class FeaturestoreService {
   @Path("/query")
   public QueryConstructorService constructQuery() {
     return queryConstructorService;
+  }
+
+  /**
+   * Verify that the name was provided as a path param
+   *
+   * @param featureStoreName the feature store name to verify
+   */
+  private void verifyNameProvided(String featureStoreName) {
+    if (Strings.isNullOrEmpty(featureStoreName)) {
+      throw new IllegalArgumentException(RESTCodes.FeaturestoreErrorCode.FEATURESTORE_NAME_NOT_PROVIDED.getMessage());
+    }
   }
 }
